# docs/swagger.yaml
basePath: /
definitions:
  dto.CarAction:
    enum:
    - forward
    - backward
    - left
    - right
    - stop
    type: string
    x-enum-varnames:
    - CarActionForward
    - CarActionBackward
    - CarActionLeft
    - CarActionRight
    - CarActionStop
  dto.CarActionRequest:
    properties:
      action:
        $ref: '#/definitions/dto.CarAction'
    type: object
  dto.GetCarsListRequest:
    properties:
      order:
        description: asc | desc
        type: string
      page:
        type: integer
      per_page:
        type: integer
      search:
        type: string
      sort_by:
        type: string
    type: object
  dto.GetCarsListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Car'
        type: array
      page:
        $ref: '#/definitions/dto.PageInfo'
    type: object
  dto.PageInfo:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  models.Car:
    properties:
      created_at:
        type: string
      id:
        type: string
      last_seen:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: SmartCar Backend API
  title: SmartCar API
  version: "1.0"
paths:
  /api/car-actions:
    post:
      consumes:
      - application/json
      description: Send action to the car via MQTT
      parameters:
      - description: Car action
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CarActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Send car action
      tags:
      - Car
  /api/cars/{id}/info:
    get:
      description: Returns information about a specific car by its ID
      parameters:
      - description: Car ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Car'
        "400":
          description: Invalid car ID
          schema:
            type: string
        "404":
          description: Car not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get car information
      tags:
      - Car
  /api/cars/{id}/stream:
    get:
      description: Sends a command to the specified car to start the camera stream
      parameters:
      - description: Car ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid car ID
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Start car camera stream
      tags:
      - Car
  /api/cars/list:
    post:
      consumes:
      - application/json
      description: Returns a paginated list of registered cars
      parameters:
      - description: Pagination parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GetCarsListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetCarsListResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get paginated list of cars
      tags:
      - Car
swagger: "2.0"